requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0